Move simulation solver from base simulation to PDE simulation
